Added fixes courtesy of Vitaly Lavrov <vel21ripn@gmail.com>
1.
This patch disables the built-in determining the direction of packet transmission.
For this I added the flag "direction_detect_disable" in the structure "ndpi_detection_module_struct",
and the flag "packet_direction" in the structure ndpi_flow_struct.

2. 
This patch reduces memory usage. To reduce the size of the structure "ndpi_packet_struct" we can delete array
"unux_line" and made the necessary changes in three protocols (icecast, irc, yahoo) who used it. It is possible
to reduce the size of the structure is about 40%.




git-svn-id: https://svn.ntop.org/svn/ntop/trunk/nDPI@8803 a02cd0c1-8e76-42e1-a119-56f9641475e2