Completely reworked Tor protocol implementation to match Tor connections over SSL.
The tor protocol uses SSL to contact peers so it could be exchanged with
standard SSL. However the host names such as

- www.fgd2iwya7vinfutj5wq5we.com
- www.qbtxzhetq4s2f.com
- www.fgd2iwya7vinfutj5wq5we.net

In essence www.<name>.com|net

To do things properly we should check if host name in the certificate
exists or if the IP address of the SSL peer resolves to a name. Unfortunately
for performance reasons we can't afford to do a DNS lookup in nDPI (however apps
can do it if possible) and thus we have created a heuristic algorithm that tries
to check the host name (in the SSL certificate) to see if it looks like a
random name or a real name. We cannot use a dictionary (how can the kernel read a file?)
and this we use bigrams distribution to decide if the string looks like a
word or just random chars.



git-svn-id: https://svn.ntop.org/svn/ntop/trunk/nDPI@8961 a02cd0c1-8e76-42e1-a119-56f9641475e2